Refactor to dedupe getDocumentRoot() methods
